requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
psycopg2-binary
azure-core==1.26.4
azure-storage-blob==12.16.0
azure-storage-file-datalake==12.11.0
